// src/types/resumeSchema.ts
export type Iso8601 = string

export interface ResumeSchema {
// ...
	volunteer?: {
		/**
		 * e.g. Facebook
		 */
		organization?: string
		/**
		 * e.g. Software Engineer
		 */
		position?: string
		/**
		 * e.g. http://facebook.example.com
		 */
		url?: string
		startDate?: Iso8601
		endDate?: Iso8601
		/**
		 * Give an overview of your responsibilities at the company
		 */
		summary?: string
		/**
		 * Specify accomplishments and achievements
		 */
		highlights?: string[]
		[k: string]: unknown
	}[]
// ...
}